Replace `sax` with `@rgrove/parse-xml`
Closes GH-3.

This commit switches from `sax`, a lax parser, to `parse-xml`, a proper and
fast parser.

This does mean that invalid XML can no longer be passed, so it’s a breaking
change (for example, all documents now require a root element, doctypes need to
be capitalized, and CDATA around the root element is no longer okay).
It also means that whitespace around the root element is no longer present
in the tree: XML requires stripping that.

It does mean that positional info is now much better, processing instructions
are supported, and the size is cut by a lot.